+++ packages/stylelint-config-wordpress/index.js
@@ -35,7 +35,6 @@ module.exports = {
     "media-query-list-comma-newline-after": "always-multi-line",
     "media-query-list-comma-space-after": "always-single-line",
     "media-query-list-comma-space-before": "never",
-    "media-query-parentheses-space-inside": "never",
     "no-eol-whitespace": true,
     "no-missing-eof-newline": true,
     "number-leading-zero": "always",